requirements.txt
pandas==2.2.2
numpy==1.26.4
streamlit==1.36.0
scikit-learn==1.4.2
xgboost==2.0.3
pyarrow==15.0.0
pytest==8.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
